+++ nengo_dl/tests/dummies.py
@@ -144,6 +144,7 @@ class TensorGraph(tensor_graph.TensorGraph):
 
         self.plan = plan
         self.minibatch_size = minibatch_size
+        self.seed = 0
 
         self.signals = signals.SignalDict(self.dtype, self.minibatch_size)